+++ src/test/java/io/codemine/java/postgresql/codecs/OidCodecIT.java
@@ -2,8 +2,8 @@ package io.codemine.java.postgresql.codecs;
 
 import io.codemine.java.postgresql.CodecITBase;
 
-public class OidCodecIT extends CodecITBase<Integer> {
+public class OidCodecIT extends CodecITBase<Long> {
   public OidCodecIT() {
-    super(Codec.OID, Integer.class);
+    super(Codec.OID, Long.class);
   }
 }
